+++ src/codegen.cpp
@@ -752,7 +752,12 @@ void ByteCodeGenerator::visitArgsNode(ArgsNode &node) {
 }
 
 void ByteCodeGenerator::visitApplyNode(ApplyNode &node) {
-  if (node.isMethodCall()) {
+  if (node.getAttr() == ApplyNode::MAP_NEXT_KEY) {
+    this->visit(node.getRecvNode());
+    this->emit0byteIns(OpCode::MAP_NEXT);
+  } else if (node.getAttr() == ApplyNode::MAP_NEXT_VALUE) {
+    // do nothing
+  } else if (node.isMethodCall()) {
     this->visit(node.getRecvNode());
     this->visit(node.getArgsNode());
     this->emitSourcePos(node.getPos());