+++ src/node.cpp
@@ -322,7 +322,9 @@ void ApplyNode::dump(NodeDumper &dumper) const {
   OP(INDEX)                                                                                        \
   OP(UNARY)                                                                                        \
   OP(BINARY)                                                                                       \
-  OP(ITER)
+  OP(ITER)                                                                                         \
+  OP(MAP_NEXT_KEY)                                                                                 \
+  OP(MAP_NEXT_VALUE)
 
   DUMP_ENUM(attr, EACH_ENUM);
 #undef EACH_ENUM
@@ -971,10 +973,10 @@ void ErrorNode::dump(NodeDumper &) const {}
 void EmptyNode::dump(NodeDumper &) const {} // do nothing
 
 // for node creation
-std::unique_ptr<LoopNode> createForInNode(unsigned int startPos, NameInfo &&varName,
-                                          std::unique_ptr<Node> &&exprNode,
+std::unique_ptr<LoopNode> createForInNode(unsigned int startPos, NameInfo &&keyName,
+                                          NameInfo &&valueName, std::unique_ptr<Node> &&exprNode,
                                           std::unique_ptr<BlockNode> &&blockNode) {
-  Token dummy = {startPos, 1};
+  Token dummy = exprNode->getToken();
 
   // create for-init
   auto call_iter = ApplyNode::newIter(std::move(exprNode));
@@ -988,15 +990,33 @@ std::unique_ptr<LoopNode> createForInNode(unsigned int startPos, NameInfo &&varN
   auto reset_var = std::make_unique<VarNode>(dummy, std::string(reset_var_name));
   auto call_hasNext = ApplyNode::newMethodCall(std::move(reset_var), std::string(OP_HAS_NEXT));
 
-  // create forIn-init
-  reset_var = std::make_unique<VarNode>(dummy, std::string(reset_var_name));
-  auto call_next = ApplyNode::newMethodCall(std::move(reset_var), std::string(OP_NEXT));
-  auto init_var = std::make_unique<VarDeclNode>(startPos, std::move(varName), std::move(call_next),
-                                                VarDeclNode::VAR);
+  // create for-in init
+  if (valueName) { // for k, v in <expr>
+    // value
+    reset_var = std::make_unique<VarNode>(dummy, std::string(reset_var_name));
+    auto call_next = ApplyNode::newMapNextValue(std::move(reset_var));
+    unsigned int pos = valueName.getToken().pos;
+    auto init_var = std::make_unique<VarDeclNode>(pos, std::move(valueName), std::move(call_next),
+                                                  VarDeclNode::VAR);
+    blockNode->insertNodeToFirst(std::move(init_var));
 
-  // insert init to block
-  blockNode->insertNodeToFirst(std::move(init_var));
+    // key
+    reset_var = std::make_unique<VarNode>(dummy, std::string(reset_var_name));
+    call_next = ApplyNode::newMapNextKey(std::move(reset_var));
+    pos = keyName.getToken().pos;
+    init_var = std::make_unique<VarDeclNode>(pos, std::move(keyName), std::move(call_next),
+                                             VarDeclNode::VAR);
+    blockNode->insertNodeToFirst(std::move(init_var));
+  } else { // for v in <expr>
+    reset_var = std::make_unique<VarNode>(dummy, std::string(reset_var_name));
+    auto call_next = ApplyNode::newMethodCall(std::move(reset_var), std::string(OP_NEXT));
+    unsigned int pos = keyName.getToken().pos;
+    auto init_var = std::make_unique<VarDeclNode>(pos, std::move(keyName), std::move(call_next),
+                                                  VarDeclNode::VAR);
 
+    // insert init to block
+    blockNode->insertNodeToFirst(std::move(init_var));
+  }
   return std::make_unique<LoopNode>(startPos, std::move(reset_varDecl), std::move(call_hasNext),
                                     nullptr, std::move(blockNode));
 }